fixed unlink issue

+++ scripts/update-server.js
@@ -73,7 +73,7 @@ app.post('/git', verifyPostData, (req, res) => {
     console.log('Request was signed successfully!');
     res.status(200).send('Request body was signed');
 
-    fs.unlinkSync('./out.txt');
+    fs.existsSync('./out.txt') && fs.unlinkSync('./out.txt');
     exec("cd /home/ubuntu/www/fds-blog && git pull origin master && hugo && cd public && tar -cf /tmp/fds_blog.tar . && curl -X POST -H \"Content-Type: application/x-tar\" -H \"Swarm-Index-Document: index.html\" -H \"Swarm-Error-Document: error.html\" --data-binary @/tmp/fds_blog.tar http://localhost:1633/dirs | jq -r .reference > out.txt", (error, stdout, stderr) => {
         if (error) {
             console.log(`exec error: ${error.message}`);
